requirements.txt
# Python 3.10 compatible dependencies
fastapi==0.104.1
uvicorn==0.24.0
pillow==10.1.0
numpy==1.26.2
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
python-multipart==0.0.6
jinja2==3.1.2
werkzeug==3.0.1

# Base dependencies
setuptools==69.0.3
wheel==0.42.0

# Computer Vision & ML libraries
opencv-python==4.8.1.78
opencv-python-headless==4.11.0.86
scikit-learn==1.3.2
pydantic==2.5.2
pydantic-settings==2.0.0

# Derm Foundation
transformers==4.36.2
huggingface-hub==0.20.3
torch==2.1.2
torchvision==0.16.2
safetensors>=0.3.1
tokenizers<0.19,>=0.14

# Testing
pytest==7.3.1

# Utilidades
python-slugify>=8.0.1
pyyaml>=6.0
httpx>=0.24.0

# Desarrollo
black>=23.3.0
flake8>=6.0.0
isort>=5.12.0